--- lib/redmine_custom_fields_sections/project_custom_field_patch.rb
@@ -1,7 +1,0 @@
-require_dependency "project_custom_field"
-
-class ProjectCustomField
-  belongs_to :section, class_name: "CustomFieldsSection", optional: true
-
-  safe_attributes("section_id")
-end
